Minor fixes from Janus
