+++ metisp/pymetis/src/pymetis/instruments/micado/readme.md
@@ -1,1 +1,1 @@
-This module will contain the shadow pipeline for MICADO.:q
+This module will contain the shadow pipeline for MICADO, if we want to do that.
